+++ src/plumbing/hash.rs
@@ -8,7 +8,7 @@ pub const SIZE: u16 = 20;
 // HexSize defines the strings size of the hash when represented in hexadecimal.
 pub const HEX_SIZE: u16 = 40;
 
-pub fn compute_hash(t: &ObjectType, content: Vec<u8>) -> String {
+pub fn compute_hash(t: &ObjectType, content: &[u8]) -> String {
     let mut hasher = Sha1::new();
     hasher.update(object_type_bytes(t));
     hasher.update(b" ");